+++ processor/src/test/java/arez/processor/ArezProcessorTest.java
@@ -207,6 +207,8 @@ public class ArezProcessorTest
         new Object[]{ "com.example.observable.SetterAlwaysMutatesFalseObjectValue", false, false, false },
         new Object[]{ "com.example.observable.SetterAlwaysMutatesFalsePrimitiveValue", false, false, false },
         new Object[]{ "com.example.observable.WildcardGenericObservableModel", false, false, false },
+        new Object[]{ "com.example.observable.WriteOutsideTransactionObservablesModel", false, false, false },
+        new Object[]{ "com.example.observable.WriteOutsideTransactionThrowingObservablesModel", false, false, false },
         new Object[]{ "com.example.observable.AbstractNonPrimitiveNonnullObservablesModel", false, false, false },
         new Object[]{ "com.example.observable.AbstractPrimitiveObservablesWithInitializerModel", false, false, false },
         new Object[]{ "com.example.observable_value_ref.DefaultRefNameModel", false, false, false },