+++ processor/src/test/resources/expected/Arez_ImplicitSingletonModel.java
@@ -133,6 +133,10 @@ public final class Arez_ImplicitSingletonModel extends ImplicitSingletonModel im
 
   @Override
   public void setTime(final long time) {
+    this.$$arezi$$_setTime( time );
+  }
+
+  private void $$arezi$$_setTime(final long time) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setTime' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }