+++ processor/src/test/resources/expected/Arez_ObservableTypeParametersModel.java
@@ -124,6 +124,10 @@ public final class Arez_ObservableTypeParametersModel<T extends Integer> extends
 
   @Override
   public void setTime(final T time) {
+    this.$$arezi$$_setTime( time );
+  }
+
+  private void $$arezi$$_setTime(final T time) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setTime' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }