+++ processor/src/test/resources/expected/com/example/deprecated/Arez_DeprecatedObservableModel1.java
@@ -126,8 +126,12 @@ public final class Arez_DeprecatedObservableModel1 extends DeprecatedObservableM
   }
 
   @Override
-  @SuppressWarnings("deprecation")
   void setTime(final long time) {
+    this.$$arezi$$_setTime( time );
+  }
+
+  @SuppressWarnings("deprecation")
+  private void $$arezi$$_setTime(final long time) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setTime' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }