+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithCreateOnly.java
@@ -142,6 +142,10 @@ final class Arez_RepositoryWithCreateOnly extends RepositoryWithCreateOnly imple
 
   @Override
   public void setName(@Nonnull final String name) {
+    this.$$arezi$$_setName( name );
+  }
+
+  private void $$arezi$$_setName(@Nonnull final String name) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setName' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }