+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithInitializerModel.java
@@ -143,6 +143,10 @@ public final class Arez_RepositoryWithInitializerModel extends RepositoryWithIni
 
   @Override
   public void setTime(final long value) {
+    this.$$arezi$$_setTime( value );
+  }
+
+  private void $$arezi$$_setTime(final long value) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setTime' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }