+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithMultipleInitializersModel.java
@@ -151,6 +151,10 @@ public final class Arez_RepositoryWithMultipleInitializersModel extends Reposito
 
   @Override
   public void setTime(final long value) {
+    this.$$arezi$$_setTime( value );
+  }
+
+  private void $$arezi$$_setTime(final long value) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setTime' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }
@@ -173,6 +177,10 @@ public final class Arez_RepositoryWithMultipleInitializersModel extends Reposito
 
   @Override
   public void setValue(final long value) {
+    this.$$arezi$$_setValue( value );
+  }
+
+  private void $$arezi$$_setValue(final long value) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setValue' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }