+++ processor/src/test/resources/expected/com/example/type_access_levels/Arez_ReduceAccessLevelModel.java
@@ -125,6 +125,10 @@ final class Arez_ReduceAccessLevelModel extends ReduceAccessLevelModel implement
 
   @Override
   public void setTime(final long time) {
+    this.$$arezi$$_setTime( time );
+  }
+
+  private void $$arezi$$_setTime(final long time) {
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setTime' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }